requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.4.0
pymongo==4.8.0
python-dotenv==1.0.1
gunicorn==21.2.0